+++ packages/gatsby-plugin-google-analytics/src/gatsby-node.js
@@ -1,9 +1,11 @@
 exports.pluginOptionsSchema = ({ Joi }) =>
   // TODO: make sure that trackingId gets required() when releasing a major version
   Joi.object({
-    trackingId: Joi.string().description(
-      `The property ID; the tracking code won't be generated without it`
-    ),
+    trackingId: Joi.string()
+      .description(
+        `The property ID; the tracking code won't be generated without it`
+      )
+      .required(),
     head: Joi.boolean()
       .default(false)
       .description(