+++ packages/gatsby-plugin-image/src/index.ts
@@ -9,12 +9,16 @@ export { StaticImage } from "./components/static-image.server"
 export {
   getImage,
   getSrc,
-  useGatsbyImage,
-  useArtDirection,
+  getSrcSet,
+  getImageData,
+  withArtDirection,
   IArtDirectedImage,
+  IGetImageDataArgs,
+  IUrlBuilderArgs,
 } from "./components/hooks"
 export {
   generateImageData,
+  getLowResolutionImageURL,
   IGatsbyImageHelperArgs,
   ISharpGatsbyImageArgs,
   IImage,