+++ packages/gatsby-transformer-remark/src/__tests__/extend-node.js
@@ -4,7 +4,7 @@ const extendNodeType = require(`../extend-node-type`)
 const { createContentDigest } = require(`gatsby-core-utils`)
 const { typeDefs } = require(`../create-schema-customization`)
 
-jest.mock(`gatsby-cli/lib/reporter`, () => {
+jest.mock(`gatsby/reporter`, () => {
   return {
     log: jest.fn(),
     info: jest.fn(),
@@ -26,14 +26,16 @@ async function queryResult(
   fragment,
   { additionalParameters = {}, pluginOptions = {} }
 ) {
+  const cache = {
+    // GatsbyCache
+    get: async () => null,
+    set: async () => null,
+  }
   const extendNodeTypeFields = await extendNodeType(
     {
       type: { name: `MarkdownRemark` },
-      cache: {
-        // GatsbyCache
-        get: async () => null,
-        set: async () => null,
-      },
+      cache,
+      getCache: () => cache,
       getNodesByType: type => [],
       ...additionalParameters,
     },
@@ -935,7 +937,7 @@ some text
 
 some other text
 `,
-    `tableOfContents
+    `tableOfContents(absolute: true, pathToSlugField: "fields.slug")
     frontmatter {
         title
     }`,
@@ -963,7 +965,7 @@ some other text
 
 final text
 `,
-    `tableOfContents(pathToSlugField: "frontmatter.title")
+    `tableOfContents(pathToSlugField: "frontmatter.title", absolute: true)
     frontmatter {
         title
     }`,
@@ -990,7 +992,7 @@ some other text
 
 final text
 `,
-    `tableOfContents(absolute: false)
+    `tableOfContents
     frontmatter {
         title
     }`,
@@ -1012,7 +1014,7 @@ some text
 ## second title
 
 some other text`,
-    `tableOfContents(pathToSlugField: "frontmatter.title", maxDepth: 1)
+    `tableOfContents(pathToSlugField: "frontmatter.title", maxDepth: 1, absolute: true)
     frontmatter {
         title
     }`,
@@ -1036,7 +1038,7 @@ some text
 ## second title
 
 some other text`,
-    `tableOfContents(pathToSlugField: "frontmatter.title")
+    `tableOfContents(pathToSlugField: "frontmatter.title", absolute: true)
     frontmatter {
         title
     }`,
@@ -1071,7 +1073,7 @@ some other text
 # third title
 
 final text`,
-    `tableOfContents(pathToSlugField: "frontmatter.title", heading: "first title")
+    `tableOfContents(pathToSlugField: "frontmatter.title", heading: "first title", absolute: true)
     frontmatter {
         title
     }`,
@@ -1101,7 +1103,7 @@ Content - content
 ### Embedding \`<code>\` Tags
 
 It's easier than you may imagine`,
-    `tableOfContents(pathToSlugField: "frontmatter.title")
+    `tableOfContents(pathToSlugField: "frontmatter.title", absolute: true)
     frontmatter {
         title
     }`,
@@ -1125,7 +1127,7 @@ Content - content
 ### Embedding \`<code>\` Tags
 
 It's easier than you may imagine`,
-    `tableOfContents(pathToSlugField: "frontmatter.title")
+    `tableOfContents(pathToSlugField: "frontmatter.title", absolute: true)
     frontmatter {
         title
     }`,